az-bridge/requirements.txt
paho-mqtt
azure-iot-device
asyncio
uuid